requirements.txt
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

chromadb>=0.5.0

dashscope>=1.20.0

streamlit>=1.36.0

pandas>=2.0.0
numpy>=1.24.0

pydantic>=2.7.0

tqdm>=4.66.0

PyYAML>=6.0

python-dotenv>=1.0.1

pypdf>=4.2.0